Break out main and create a core sub module for running the main logic of cfnlint
