+++ src/cfnlint/cfn_yaml.py
