+++ src/cfnlint/helpers.py
@@ -22,10 +22,7 @@ import imp
 import logging
 import re
 import requests
-from yaml.parser import ParserError, ScannerError
 import pkg_resources
-from cfnlint.parser import DuplicateError, NullError
-import cfnlint.parser
 
 
 LOGGER = logging.getLogger(__name__)
@@ -243,95 +240,3 @@ def override_specs(override_spec_file):
     except (ValueError) as err:
         LOGGER.error('Override spec file %s is malformed: %s', filename, err)
         sys.exit(1)
-
-
-def get_template_default_args(filename, ignore_bad_template):
-    """ Get Template Configuration items and set them as default vales"""
-    defaults = {}
-    template = {}
-    try:
-        fp = open(filename)
-        loader = cfnlint.parser.MarkedLoader(fp.read())
-        loader.add_multi_constructor('!', cfnlint.parser.multi_constructor)
-        template = loader.get_single_data()
-        if isinstance(template, dict):
-            defaults = template.get('Metadata', {}).get('cfn-lint', {}).get('config', {})
-    except IOError as e:
-        if e.errno == 2:
-            LOGGER.error('Template file not found: %s', filename)
-            sys.exit(1)
-        elif e.errno == 21:
-            LOGGER.error('Template references a directory, not a file: %s', filename)
-            sys.exit(1)
-        elif e.errno == 13:
-            LOGGER.error('Permission denied when accessing template file: %s', filename)
-            sys.exit(1)
-    except DuplicateError as err:
-        LOGGER.error('Template %s contains duplicates: %s', filename, err)
-        sys.exit(1)
-    except NullError as err:
-        LOGGER.error('Template %s contains nulls: %s', filename, err)
-        sys.exit(1)
-    except (ParserError, ScannerError) as err:
-        try:
-            template = json.load(open(filename), cls=cfnlint.cfn_json.CfnJSONDecoder)
-        except cfnlint.cfn_json.JSONDecodeError as json_err:
-            if ignore_bad_template:
-                LOGGER.info('Template %s is malformed: %s', filename, err.problem)
-                LOGGER.error('Tried to parse %s as JSON but got error: %s', filename, str(json_err))
-            else:
-                LOGGER.error('Template %s is malformed: %s', filename, err.problem)
-                LOGGER.error('Tried to parse %s as JSON but got error: %s', filename, str(json_err))
-            sys.exit(1)
-        except Exception as json_err:  # pylint: disable=W0703
-            if ignore_bad_template:
-                LOGGER.info('Template %s is malformed: %s', filename, err.problem)
-                LOGGER.info('Tried to parse %s as JSON but got error: %s', filename, str(json_err))
-            else:
-                LOGGER.error('Template %s is malformed: %s', filename, err.problem)
-                LOGGER.error('Tried to parse %s as JSON but got error: %s', filename, str(json_err))
-                sys.exit(1)
-
-    return(defaults, template)
-
-
-def run_checks(filename, template, rules, transforms, ignore_checks, regions):
-    """Run Checks against the template"""
-    if regions:
-        supported_regions = [
-            'ap-south-1',
-            'sa-east-1',
-            'ap-northeast-1',
-            'ap-northeast-2',
-            'ap-southeast-1',
-            'ap-southeast-2',
-            'ca-central-1',
-            'eu-central-1',
-            'eu-west-1',
-            'eu-west-2',
-            'us-west-2',
-            'us-east-1',
-            'us-east-2',
-            'us-west-1'
-        ]
-        for region in regions:
-            if region not in supported_regions:
-                LOGGER.error('Supported regions are %s', supported_regions)
-                exit(32)
-
-    matches = list()
-
-    runner = cfnlint.Runner(
-        rules, transforms, filename, template,
-        ignore_checks, regions)
-    matches.extend(runner.transform())
-    # Only do rule analysis if Transform was successful
-    if not matches:
-        try:
-            matches.extend(runner.run())
-        except Exception as err:  # pylint: disable=W0703
-            LOGGER.error('Tried to process rules on file %s but got an error: %s', filename, str(err))
-            exit(1)
-    matches.sort(key=lambda x: (x.filename, x.linenumber, x.rule.id))
-
-    return(matches)
