+++ test/module/core/__init__.py
