+++ test/module/test_null_values.py
@@ -15,8 +15,9 @@
   SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
 import json
-from cfnlint import RulesCollection, DEFAULT_RULESDIR  # pylint: disable=E0401
-import cfnlint.parser  # pylint: disable=E0401
+from cfnlint import RulesCollection  # pylint: disable=E0401
+from cfnlint.core import DEFAULT_RULESDIR
+import cfnlint.cfn_yaml  # pylint: disable=E0401
 import cfnlint.cfn_json  # pylint: disable=E0401
 from testlib.testcase import BaseTestCase
 
@@ -38,10 +39,10 @@ class TestNulls(BaseTestCase):
 
         try:
             fp = open(filename)
-            loader = cfnlint.parser.MarkedLoader(fp.read())
-            loader.add_multi_constructor('!', cfnlint.parser.multi_constructor)
+            loader = cfnlint.cfn_yaml.MarkedLoader(fp.read())
+            loader.add_multi_constructor('!', cfnlint.cfn_yaml.multi_constructor)
             loader.get_single_data()
-        except cfnlint.parser.NullError:
+        except cfnlint.cfn_yaml.NullError:
             assert(False)
             return
 
@@ -70,10 +71,10 @@ class TestNulls(BaseTestCase):
 
         try:
             fp = open(filename)
-            loader = cfnlint.parser.MarkedLoader(fp.read())
-            loader.add_multi_constructor('!', cfnlint.parser.multi_constructor)
+            loader = cfnlint.cfn_yaml.MarkedLoader(fp.read())
+            loader.add_multi_constructor('!', cfnlint.cfn_yaml.multi_constructor)
             loader.get_single_data()
-        except cfnlint.parser.NullError:
+        except cfnlint.cfn_yaml.NullError:
             assert(True)
             return
