+++ test/module/test_rules_collections.py
@@ -14,8 +14,9 @@
   OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
   SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
-from cfnlint import RulesCollection, Template, DEFAULT_RULESDIR  # pylint: disable=E0401
-import cfnlint.parser  # pylint: disable=E0401
+from cfnlint import RulesCollection, Template
+from cfnlint.core import DEFAULT_RULESDIR  # pylint: disable=E0401
+import cfnlint.cfn_yaml  # pylint: disable=E0401
 from testlib.testcase import BaseTestCase
 
 
@@ -40,8 +41,8 @@ class TestTemplate(BaseTestCase):
         """ Test Run Logic"""
         filename = 'templates/good/generic.yaml'
         fp = open(filename)
-        loader = cfnlint.parser.MarkedLoader(fp.read())
-        loader.add_multi_constructor("!", cfnlint.parser.multi_constructor)
+        loader = cfnlint.cfn_yaml.MarkedLoader(fp.read())
+        loader.add_multi_constructor("!", cfnlint.cfn_yaml.multi_constructor)
         template = loader.get_single_data()
         cfn = Template(template, ['us-east-1'])
 
@@ -53,8 +54,8 @@ class TestTemplate(BaseTestCase):
         """Test failure run"""
         filename = 'templates/bad/generic.yaml'
         fp = open(filename)
-        loader = cfnlint.parser.MarkedLoader(fp.read())
-        loader.add_multi_constructor("!", cfnlint.parser.multi_constructor)
+        loader = cfnlint.cfn_yaml.MarkedLoader(fp.read())
+        loader.add_multi_constructor("!", cfnlint.cfn_yaml.multi_constructor)
         template = loader.get_single_data()
         cfn = Template(template, ['us-east-1'])
 
@@ -66,8 +67,8 @@ class TestTemplate(BaseTestCase):
         """Test failure run"""
         filename = 'templates/bad/properties_onlyone.yaml'
         fp = open(filename)
-        loader = cfnlint.parser.MarkedLoader(fp.read())
-        loader.add_multi_constructor("!", cfnlint.parser.multi_constructor)
+        loader = cfnlint.cfn_yaml.MarkedLoader(fp.read())
+        loader.add_multi_constructor("!", cfnlint.cfn_yaml.multi_constructor)
         template = loader.get_single_data()
         cfn = Template(template, ['us-east-1'])
