+++ test/templates/bad/core/config_invalid_json.json
