+++ test/templates/bad/core/config_invalid_yaml.yaml
