+++ test/templates/bad/core/config_parameters.yaml
@@ -1,0 +1,7 @@
+---
+AWSTemplateFormatVersion: "2010-09-09"
+Metadata:
+  cfn-lint:
+    config:
+      ignore_checks: E0101
+Resources: {}
