+++ test/templates/good/core/config_cfn_lint.json
